requirements.txt
# Student Attendance System - Requirements

# Core web framework
streamlit>=1.28.0

# Computer vision and machine learning
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0

# Utilities
PyYAML>=6.0
pathlib2>=2.3.0

# Optional: For GPU support (CUDA)
# torch>=2.0.0+cu118
# torchvision>=0.15.0+cu118